+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 
+## [0.2.0] - 2015-05-03
+
 ### Added
 
 - Support HTML string when searching for attributes with Floki.attribute/2.
@@ -44,7 +46,8 @@ of the parent element inside HTML.
 
 - Elixir version requirement from "~> 1.0.0" to ">= 1.0.0".
 
-[unreleased]: https://github.com/philss/floki/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/philss/floki/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/philss/floki/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/philss/floki/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/philss/floki/compare/v0.0.5...v0.1.0
 [0.0.5]: https://github.com/philss/floki/compare/v0.0.3...v0.0.5